maked component to confirm actions

+++ components/data-table.tsx
@@ -24,6 +24,7 @@ import {
 import { Button } from "./ui/button";
 import { Input } from "./ui/input";
 import { Trash } from "lucide-react";
+import { useConfirm } from "@/hooks/use-confirm";
 
 interface DataTableProps<TData, TValue> {
     columns: ColumnDef<TData, TValue>[];
@@ -40,6 +41,10 @@ export function DataTable<TData, TValue>({
     onDelete,
     disabled,
 }: DataTableProps<TData, TValue>) {
+    const [ConfirmDialog, confirm] = useConfirm(
+        "Are you sure?",
+        "You are about to perform a bulk delete"
+    );
     const [sorting, setSorting] = React.useState<SortingState>([]);
     const [columnFilters, setColumnFilters] =
         React.useState<ColumnFiltersState>([]);
@@ -65,6 +70,7 @@ export function DataTable<TData, TValue>({
 
     return (
         <div>
+            <ConfirmDialog />
             <div className="flex items-center py-4">
                 <Input
                     placeholder={`Filter ${filterKey}`}
@@ -85,10 +91,14 @@ export function DataTable<TData, TValue>({
                     size="sm"
                     variant="outline"
                     className="ml-auto font-normal text-xs"
-                    disabled={disabled}
-                    onClick={() => {
-                        onDelete(table.getFilteredSelectedRowModel().rows);
-                        table.resetRowSelection();
+                    disabled={disabled || !table.getFilteredSelectedRowModel().rows.length}
+                    onClick={async () => {
+                        const ok = await confirm()
+
+                        if(ok) {
+                            onDelete(table.getFilteredSelectedRowModel().rows);
+                            table.resetRowSelection();
+                        }
                     }}
                 >
                     <Trash className="size-4 mr-2" />
